Add top level menu to WordPress admin menu.
Currently nothing useful on the page, but a placeholder for a more
useful page later, and possibly submenus for other functionality.

The menu is placed right below the 'Media' menu for now, and is
available to anyone with the 'upload_files' capability. Both of these
may be adjusted later if there's a better location or other capabilities
fit better.