Refactor processing uploads.
Moves all database operations to separate classes for each type of
object, bands, venues and concerts. This makes the logic cleaner, and
concentrates the db related code to logically distinct parts of the
code.

This means we no longer need the global access to the $wpdb object from
the processing code.