Populate the pressstatus table on activation.
Was missing in the first import. This looks like pure static data,
though, and will probably be moved to constants in the code instead.

While this doesn't really change the db schema, the db version was
incremented to version 2, to make sure the new data is added, even if
the db is already initialized.