Change venues into proper objects.
This means most static functions now either return a venue object, or an
array of venue objects. The exception is the `all_cities` method, which
still return an array of cities as strings.

The constructor has been made private, as it should not be used directly
from anywhere but the static methods on the Venue class.