Fix test setup for wp-env.
After much reading I finally found the magic incantations, so now we can
run tests with real database access. This means we no longer need the
primitive $wpdb_stub.

The setup as now _requires_ wp-env, or an environment set up
sufficiently similar. Running in wp-env is the easiest, so aim for that.

I've added a `run-tests` script that will invoke the magic incantation
without having to remember it every time.

To set up for testing:

1. make sure you have composer[1] installed.
2. run `composer install`
3. make sure you have wp-env[2] installed
4. start the wordpress env: `wp-env start`
5. run the tests: `./run-tests`

Let the thousand tests bloom!

[1]: https://github.com/wp-phpunit/wp-phpunit
[2]: https://www.npmjs.com/package/@wordpress/env