Refactor Concert::create
Restructure the method a bit, drop the `c`prefix on variables, rename
the variable to hold the created concert, and use object notation to
pass the attributes to the constructor.

Also rename the method `get` to `find`, the only call site was the
`create` method. Drop the unnecessary method `check_duplicate`. Just use
`find` instead.