Include info from venue in concerts.
This makes the concert a full object containing all relevant info, while
we can still segment the data in the db.

Instead of this:

    $concert = GiglogAdmin_Concert::get($concert_id);
    $venue = GiglogAdmin_Venue::get($concert->venue());
    echo "{$concert->name()} @ {$venue->name()} : {$concert->cdate()}"

You can now do:

    $concert = GiglogAdmin_Concert::get($concert_id);
    echo "{$concert->name()} @ {$concert->venue()->name()} : {$concert->cdate()}"

And yeah, renamed Concert::find_cid() to Concert::get() and changed it's
semantics somewhat. It now either returns the given concert if it
exists, or NULL if it does not. Simpler function; simpler to use.