Redo concert list to use the Concert class.
Instead of directly accessing the database with a custom query, we now
just use the Concert::find_concerts method to fetch the concerts that
are to be displayed.

This became much easier now that we don't rely on the extra concertlogs
table.

There's still stuff missing to be functionally equivalent to the old
code, but this should be a lot easier to get in place now.