Fix so concert id is not overwritten by venue id.
We originally had a more specified query, but simplified it to:

   SELECT * FROM wpg_concerts LEFT JOIN wpg_venues ON ...;

But since both the concerts table and the venues table has a column id,
the concert id would be overwritten with the venue id. MySQL/MariaDB
does not allow columns with the same name in multiple tables when using
unqualified column names in the query.

So we need to be more explicit again.

I was hoping that the following would work:

    SELECT wpg_concerts.*, wpg_venues.* FROM .... ;

I think MySQL/MariaDB would handle that, but now since php turns the
result into an array, where each key must be unique, this again
overwrites the concert id with the venue id.

So thus a more verbose specification of the columns was necessary.