Move the concerts table and edit form out of AdminPage.
Currently the AdminPage is still responsible for updating changes to any
of the concerts, but I'd like to get that into their respective classes
too. That way the AdminPage will just be a simple class to handle the
layout of the page, while all the specific functionality is in their
own classes.

This is also the first step to be able to reuse the concerts table on
the public end of the site.