tests: Make an actual test for ConcertsTable
There's a bit of setup to make this work as it should, we need to ensure
that the current user and current screen is set to proper values so that
the WordPress api's `is_admin()` and `current_user_can()` work as they
should.

This first test just tests that all the expected forms are being
rendered for the admin user accessing the table through the site admin
interface.